JavaScript code changed to TypeScript code.
