--- models/Address.js
@@ -1,16 +1,0 @@
-const mongoose = require("mongoose");
-
-const addressModel = new mongoose.Schema(
-  {
-    user: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
-    line1: { type: String, required: true },
-    line2: { type: String },
-    city: { type: String, required: true },
-    state: { type: String, required: true },
-    zipCode: { type: String, required: true },
-    country: { type: String, required: true },
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("Address", addressModel);
