--- models/LocationModel.js
@@ -1,14 +1,0 @@
-const mongoose = require("mongoose");
-
-const locationSchema = new mongoose.Schema(
-  {
-    address: { type: String, required: true },
-    latitude: { type: Number },
-    longitude: { type: Number },
-  },
-  {
-    timestamps: true,
-  }
-);
-
-module.exports = mongoose.model("Location", locationSchema);
