--- models/Review.js
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-const reviewSchema = new mongoose.Schema(
-  {
-    user: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
-    restaurant: { type: mongoose.Schema.Types.ObjectId, ref: "Restaurant" },
-    rating: { type: Number, required: true },
-    comment: { type: String, required: true },
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("Review", reviewSchema);
