+++ models/index.ts
@@ -1,0 +1,8 @@
+export * from "./Address";
+export * from "./Favorite";
+export * from "./Location";
+export * from "./MenuItem";
+export * from "./Order";
+export * from "./Restaurant";
+export * from "./Review";
+export * from "./User";
