+++ package.json
@@ -2,10 +2,11 @@
   "name": "restaurant-api",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index"
+    "start": "nodemon dist/index.js",
+    "build": "tsc -p tsconfig.json",
+    "dev": "nodemon --exec ts-node index.ts"
   },
   "author": "Naveen",
   "license": "ISC",
@@ -18,5 +19,13 @@
     "mongoose": "^7.5.2",
     "multer": "^1.4.5-lts.1",
     "nodemon": "^3.0.1"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.14",
+    "@types/crypto-js": "^4.1.2",
+    "@types/express": "^4.17.18",
+    "@types/jsonwebtoken": "^9.0.3",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.2.2"
   }
 }
