+++ routes/index.ts
@@ -1,0 +1,4 @@
+export * from "./auth";
+export * from "./user";
+export * from "./restaurant";
+export * from "./menuItem";
